--- a/system-parachains/bridge-hubs/bridge-hub-kusama/primitives/src/lib.rs
+++ b/system-parachains/bridge-hubs/bridge-hub-kusama/primitives/src/lib.rs
@@ -98,7 +98,7 @@ frame_support::parameter_types! {
 	/// The XCM fee that is paid for executing XCM program (with `ExportMessage` instruction) at the Kusama
 	/// BridgeHub.
 	/// (initially was calculated by test `BridgeHubKusama::can_calculate_weight_for_paid_export_message_with_reserve_transfer` + `33%`)
-	pub const BridgeHubKusamaBaseXcmFeeInKsms: u128 = 386_365_000;
+	pub const BridgeHubKusamaBaseXcmFeeInKsms: u128 = 590_387_000;
 
 	/// Transaction fee that is paid at the Kusama BridgeHub for delivering single inbound message.
 	/// (initially was calculated by test `BridgeHubKusama::can_calculate_fee_for_complex_message_delivery_transaction` + `33%`)