--- a/system-parachains/bridge-hubs/bridge-hub-polkadot/primitives/src/lib.rs
+++ b/system-parachains/bridge-hubs/bridge-hub-polkadot/primitives/src/lib.rs
@@ -89,7 +89,7 @@ frame_support::parameter_types! {
 	/// The XCM fee that is paid for executing XCM program (with `ExportMessage` instruction) at the Polkadot
 	/// BridgeHub.
 	/// (initially was calculated by test `BridgeHubPolkadot::can_calculate_weight_for_paid_export_message_with_reserve_transfer` + `33%`)
-	pub const BridgeHubPolkadotBaseXcmFeeInDots: Balance = 115_909_500;
+	pub const BridgeHubPolkadotBaseXcmFeeInDots: Balance = 177_594_900;
 
 	/// Transaction fee that is paid at the Polkadot BridgeHub for delivering single inbound message.
 	/// (initially was calculated by test `BridgeHubPolkadot::can_calculate_fee_for_complex_message_delivery_transaction` + `33%`)