Relates to: https://github.com/polkadot-fellows/runtimes/issues/140

Chain-specs generated from the commit: `f5aeb7e8cc024a17a42ca547645f25f2ae28e4f4`.
Polkadot binary used:

```
$ git checkout origin/release-polkadot-v1.7.2
$ cargo build --profile production --features runtime-benchmarks --bin polkadot
$ ./target/production/polkadot -V
polkadot 1.7.2-0d7ca59fd4b
```

<!--
```
branislav@ggwpez-ref-hw:~$ git clone https://github.com/paritytech/polkadot-sdk
branislav@ggwpez-ref-hw:~$ cd polkadot-sdk/
branislav@ggwpez-ref-hw:~/polkadot-sdk$ git checkout polkadot-v1.8.0
Note: switching to 'polkadot-v1.8.0'.
HEAD is now at ec7817e5adc adjust release notifications GHA
branislav@ggwpez-ref-hw:~/polkadot-sdk$ cargo build --profile production --features runtime-benchmarks --bin polkadot
branislav@ggwpez-ref-hw:~/polkadot-sdk$ ./target/production/polkadot -V
```
-->

## TODO

- [x] fix bug for Kusama:
   ```
   [pallet_balances, Nis] // as alias for `type Nis = pallet_balances::Pallet::<Runtime, NisCounterpartInstance>;`
   [pallet_nis, Nis]
   ```
- [x] add comparisions
  - [x] add `pallet_xcm_benchmarks_fungible.rs` to diffs - https://github.com/ggwpez/substrate-weight-compare/issues/60
- [ ] check if to use `polkadot` binary from "lastest release- v1.8.0  "  or "latest compatible bump release - v1.7.2" ?
- [x] investigate and review/approve
  - [x]  `set_candidacy_bond`  https://github.com/polkadot-fellows/runtimes/issues/230
  - [x] investigate `teleport` / `reserver_assets_transfer` (they were not regenerated with v1.1.2 + compare with regenerated in polkadot-sdk) https://github.com/polkadot-fellows/runtimes/issues/231


## The comparison of this branch with `release-v1.1.2` (latest release)

_Note: `Change [%]` has threashold 5, so it show only if change is more than `5%` or less than `-5%`._

### Polkadot weights

```
subweight compare commits \
         --path-pattern "./relay/polkadot/**/weights/**/*.rs" \
         --format markdown --no-color \
          --change added changed \
         --method asymptotic --ignore-errors \
         remotes/polkadot-fellows/release-v1.1.2 \
         origin/bko-weights 
```

| File                                                               | Extrinsic                                           | Old      | New      | Change [%] |
|--------------------------------------------------------------------|-----------------------------------------------------|----------|----------|------------|
| relay/polkadot/src/weights/pallet_xcm.rs                           | reserve_transfer_assets                             | 17.26us  | 530.08us | +2971.12   |
| relay/polkadot/src/weights/pallet_xcm.rs                           | teleport_assets                                     | 17.82us  | 534.41us | +2899.79   |
| relay/polkadot/src/weights/pallet_message_queue.rs                 | service_page_item                                   | 91.98us  | 369.58us | +301.81    |
| relay/polkadot/src/weights/pallet_xcm.rs                           | notify_target_migration_fail                        | 84.81us  | 112.31us | +32.42     |
| relay/polkadot/src/weights/runtime_parachains_disputes_slashing.rs | report_dispute_lost                                 | 1.94ms   | 2.42ms   | +24.81     |
| relay/polkadot/src/weights/pallet_election_provider_multi_phase.rs | create_snapshot_internal                            | 1.36ms   | 1.65ms   | +21.05     |
| relay/polkadot/src/weights/pallet_xcm.rs                           | already_notified_target                             | 141.46us | 169.01us | +19.48     |
| relay/polkadot/src/weights/pallet_identity.rs                      | set_identity                                        | 199.30us | 235.18us | +18.00     |
| relay/polkadot/src/weights/pallet_xcm.rs                           | migrate_supported_version                           | 314.88us | 341.71us | +8.52      |
| relay/polkadot/src/weights/pallet_xcm.rs                           | migrate_version_notify_targets                      | 315.23us | 342.07us | +8.52      |
| relay/polkadot/src/weights/pallet_xcm.rs                           | migrate_version_notifiers                           | 314.85us | 341.59us | +8.50      |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_fungible.rs   | receive_teleported_asset                            | 143.61us | 153.77us | +7.07      |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_fungible.rs   | withdraw_asset                                      | 144.31us | 153.46us | +6.34      |
| relay/polkadot/src/weights/pallet_election_provider_multi_phase.rs | submit                                              | 470.64us | 495.83us | +5.35      |
| relay/polkadot/src/weights/pallet_multisig.rs                      | as_multi_threshold_1                                | 19.02us  | 18.07us  | -5.02      |
| relay/polkadot/src/weights/runtime_common_crowdloan.rs             | poke                                                | 181.21us | 172.02us | -5.07      |
| relay/polkadot/src/weights/runtime_common_paras_registrar.rs       | set_current_head                                    | 1.13ms   | 1.08ms   | -5.11      |
| relay/polkadot/src/weights/runtime_parachains_paras.rs             | force_set_current_head                              | 1.14ms   | 1.08ms   | -5.34      |
| relay/polkadot/src/weights/pallet_staking.rs                       | bond                                                | 541.11us | 511.73us | -5.43      |
| relay/polkadot/src/weights/runtime_common_crowdloan.rs             | add_memo                                            | 192.99us | 182.39us | -5.49      |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | claim_asset                                         | 141.77us | 133.54us | -5.80      |
| relay/polkadot/src/weights/runtime_parachains_paras.rs             | include_pvf_check_statement_finalize_upgrade_reject | 298.85us | 280.69us | -6.08      |
| relay/polkadot/src/weights/pallet_conviction_voting.rs             | remove_vote                                         | 695.72us | 652.95us | -6.15      |
| relay/polkadot/src/weights/runtime_common_crowdloan.rs             | edit                                                | 153.27us | 143.05us | -6.67      |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | schedule_named                                      | 308.09us | 286.03us | -7.16      |
| relay/polkadot/src/weights/runtime_common_paras_registrar.rs       | schedule_code_upgrade                               | 7.48ms   | 6.92ms   | -7.47      |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | service_task_fetched                                | 5.72ms   | 5.23ms   | -8.58      |
| relay/polkadot/src/weights/pallet_utility.rs                       | dispatch_as                                         | 6.81us   | 6.17us   | -9.43      |
| relay/polkadot/src/weights/pallet_whitelist.rs                     | dispatch_whitelisted_call                           | 6.09ms   | 5.51ms   | -9.49      |
| relay/polkadot/src/weights/pallet_preimage.rs                      | note_requested_preimage                             | 10.66ms  | 9.65ms   | -9.52      |
| relay/polkadot/src/weights/pallet_preimage.rs                      | note_no_deposit_preimage                            | 10.67ms  | 9.65ms   | -9.56      |
| relay/polkadot/src/weights/pallet_preimage.rs                      | note_preimage                                       | 9.69ms   | 8.76ms   | -9.65      |
| relay/polkadot/src/weights/pallet_utility.rs                       | as_derivative                                       | 4.52us   | 4.06us   | -10.13     |
| relay/polkadot/src/weights/pallet_election_provider_multi_phase.rs | feasibility_check                                   | 8.56ms   | 7.63ms   | -10.89     |
| relay/polkadot/src/weights/runtime_parachains_paras_inherent.rs    | enter_backed_candidates_variable                    | 14.68ms  | 13.08ms  | -10.96     |
| relay/polkadot/src/weights/pallet_utility.rs                       | force_batch                                         | 3.44ms   | 3.05ms   | -11.42     |
| relay/polkadot/src/weights/runtime_parachains_paras.rs             | force_schedule_code_upgrade                         | 9.17ms   | 8.12ms   | -11.47     |
| relay/polkadot/src/weights/pallet_session.rs                       | purge_keys                                          | 892.66us | 790.23us | -11.47     |
| relay/polkadot/src/weights/pallet_utility.rs                       | batch                                               | 3.46ms   | 3.05ms   | -11.75     |
| relay/polkadot/src/weights/pallet_session.rs                       | set_keys                                            | 1.09ms   | 954.57us | -12.19     |
| relay/polkadot/src/weights/frame_system.rs                         | set_code                                            | 112.29ms | 98.55ms  | -12.23     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | service_task_base                                   | 3.62us   | 3.17us   | -12.42     |
| relay/polkadot/src/weights/pallet_utility.rs                       | batch_all                                           | 3.74ms   | 3.27ms   | -12.67     |
| relay/polkadot/src/weights/pallet_xcm.rs                           | force_default_xcm_version                           | 2.29us   | 2.00us   | -12.70     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | schedule                                            | 178.20us | 155.20us | -12.91     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | service_agenda_base                                 | 171.44us | 148.11us | -13.61     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | cancel_named                                        | 342.55us | 294.38us | -14.06     |
| relay/polkadot/src/weights/pallet_conviction_voting.rs             | undelegate                                          | 88.44ms  | 75.76ms  | -14.33     |
| relay/polkadot/src/weights/pallet_conviction_voting.rs             | delegate                                            | 88.76ms  | 75.94ms  | -14.44     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | cancel                                              | 314.97us | 265.63us | -15.66     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | expect_pallet                                       | 8.87us   | 7.33us   | -17.31     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | query_response                                      | 37.73us  | 30.26us  | -19.78     |
| relay/polkadot/src/weights/runtime_parachains_paras_inherent.rs    | enter_variable_disputes                             | 14.05ms  | 11.19ms  | -20.34     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | execute_dispatch_unsigned                           | 2.43us   | 1.87us   | -22.97     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | service_task_periodic                               | 4.26us   | 3.24us   | -23.94     |
| relay/polkadot/src/weights/frame_benchmarking_baseline.rs          | sr25519_verification                                | 5.49ms   | 4.09ms   | -25.46     |
| relay/polkadot/src/weights/pallet_scheduler.rs                     | execute_dispatch_signed                             | 2.52us   | 1.86us   | -26.11     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | subscribe_version                                   | 611.52us | 450.38us | -26.35     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | refund_surplus                                      | 3.46us   | 1.90us   | -45.29     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | transact                                            | 13.43us  | 7.20us   | -46.38     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | buy_execution                                       | 3.15us   | 1.44us   | -54.20     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | clear_transact_status                               | 3.10us   | 1.42us   | -54.30     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | expect_asset                                        | 3.33us   | 1.51us   | -54.47     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | expect_transact_status                              | 3.36us   | 1.53us   | -54.49     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | clear_origin                                        | 3.01us   | 1.35us   | -55.02     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | clear_topic                                         | 3.01us   | 1.34us   | -55.56     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | clear_error                                         | 3.04us   | 1.33us   | -56.14     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | set_topic                                           | 3.06us   | 1.33us   | -56.59     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | set_error_handler                                   | 3.07us   | 1.33us   | -56.65     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | set_appendix                                        | 3.11us   | 1.34us   | -56.81     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | trap                                                | 3.04us   | 1.31us   | -56.87     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | set_fees_mode                                       | 3.09us   | 1.32us   | -57.08     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | expect_error                                        | 3.13us   | 1.33us   | -57.54     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | expect_origin                                       | 3.15us   | 1.30us   | -58.61     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | unpaid_execution                                    | 3.25us   | 1.34us   | -58.63     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | descend_origin                                      | 3.90us   | 1.44us   | -63.05     |
| relay/polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs    | burn_asset                                          | 4.80us   | 1.68us   | -64.94     |
| relay/polkadot/src/weights/runtime_parachains_configuration.rs     | set_node_feature                                    |          | 184.87us | Added      |
| relay/polkadot/src/weights/pallet_xcm.rs                           | transfer_assets                                     |          | 535.72us | Added      |
| relay/polkadot/src/weights/pallet_xcm.rs                           | take_response                                       |          | 147.19us | Added      |
| relay/polkadot/src/weights/pallet_xcm.rs                           | new_query                                           |          | 227.23us | Added      |
| relay/polkadot/src/weights/pallet_xcm.rs                           | claim_assets                                        |          | 159.29us | Added      |
| relay/polkadot/src/weights/pallet_vesting.rs                       | force_remove_vesting_schedule                       |          | 440.07us | Added      |
| relay/polkadot/src/weights/pallet_staking.rs                       | update_payee                                        |          | 196.80us | Added      |
| relay/polkadot/src/weights/pallet_staking.rs                       | deprecate_controller_batch                          |          | 1.95s    | Added      |
| relay/polkadot/src/weights/pallet_nomination_pools.rs              | set_commission_claim_permission                     |          | 140.57us | Added      |
| relay/polkadot/src/weights/pallet_identity.rs                      | set_username_for                                    |          | 467.02us | Added      |
| relay/polkadot/src/weights/pallet_identity.rs                      | set_primary_username                                |          | 165.35us | Added      |
| relay/polkadot/src/weights/pallet_identity.rs                      | remove_username_authority                           |          | 134.04us | Added      |
| relay/polkadot/src/weights/pallet_identity.rs                      | remove_expired_approval                             |          | 145.93us | Added      |
| relay/polkadot/src/weights/pallet_identity.rs                      | remove_dangling_username                            |          | 160.97us | Added      |
| relay/polkadot/src/weights/pallet_identity.rs                      | add_username_authority                              |          | 106.44us | Added      |
| relay/polkadot/src/weights/pallet_identity.rs                      | accept_username                                     |          | 369.75us | Added      |
| relay/polkadot/src/weights/pallet_balances.rs                      | force_adjust_total_issuance                         |          | 5.57us   | Added      |
| relay/polkadot/src/weights/frame_system.rs                         | authorize_upgrade                                   |          | 118.02us | Added      |
| relay/polkadot/src/weights/frame_system.rs                         | apply_authorized_upgrade                            |          | 104.14ms | Added      |


### Kusama weights

```
subweight compare commits \
         --path-pattern "./relay/kusama/**/weights/**/*.rs" \
         --format markdown --no-color \
          --change added changed \
         --method asymptotic --ignore-errors \
         remotes/polkadot-fellows/release-v1.1.2 \
         origin/bko-weights 
```

| File                                                                 | Extrinsic                                              | Old      | New      | Change [%] |
|----------------------------------------------------------------------|--------------------------------------------------------|----------|----------|------------|
| relay/kusama/src/weights/pallet_xcm.rs                               | reserve_transfer_assets                                | 17.28us  | 532.85us | +2982.93   |
| relay/kusama/src/weights/pallet_xcm.rs                               | teleport_assets                                        | 17.68us  | 534.45us | +2923.79   |
| relay/kusama/src/weights/pallet_message_queue.rs                     | service_page_item                                      | 91.68us  | 369.81us | +303.38    |
| relay/kusama/src/weights/pallet_ranked_collective.rs                 | remove_member                                          | 4.40ms   | 6.62ms   | +50.40     |
| relay/kusama/src/weights/pallet_ranked_collective.rs                 | demote_member                                          | 535.45us | 737.37us | +37.71     |
| relay/kusama/src/weights/pallet_xcm.rs                               | notify_target_migration_fail                           | 84.86us  | 112.44us | +32.50     |
| relay/kusama/src/weights/runtime_parachains_disputes_slashing.rs     | report_dispute_lost                                    | 1.93ms   | 2.42ms   | +25.66     |
| relay/kusama/src/weights/pallet_xcm.rs                               | already_notified_target                                | 141.81us | 169.44us | +19.48     |
…