@ggwpez 2024-03-08
Btw  `subweight` has a `--format markdown --no-color` options 😉  
But i just noticed that it is printed to stderr instead of stdout 🙈