@bkontur 2024-03-19
> Approving assuming the weight regressions are [clarified](https://github.com/polkadot-fellows/runtimes/issues/231#issuecomment-1988755243)

actually, it looks like we decrease fees e.g. for `pallet_xcm::reserve_transfer_assets` - [more info](https://github.com/polkadot-fellows/runtimes/issues/231#issuecomment-2006851713)