@acatangiu 2024-03-19
I'm confused, shouldn't you run benchmarks again after applying fix patch?